+++ smartsheet-csharp-sdk/main/Smartsheet/Api/Internal/CommentAttachmentResourcesImpl.cs
@@ -71,6 +71,30 @@ namespace Smartsheet.Api.Internal
         }
 
         /// <summary>
+        /// <para>Attaches a file to the Comment using a Stream.</para>
+        /// <para>This operation will always create a new attachment.
+        /// To upload a new version of the same attachment, use the Attach New Version operation.</para>
+        /// <para>It mirrors to the following Smartsheet REST API method:
+        /// POST /sheets/{sheetId}/comments/{commentId}/attachments</para>
+        /// </summary>
+        /// <param name="sheetId"> the sheetId </param>
+        /// <param name="commentId"> the comment Id </param>
+        /// <param name="stream"> the file stream </param>
+        /// <param name="fileName"> the file name </param>
+        /// <param name="contentType"> the content type </param>
+        /// <returns> the newly created Attachment </returns>
+        /// <exception cref="System.InvalidOperationException"> if any argument is null or empty string </exception>
+        /// <exception cref="InvalidRequestException"> if there is any problem with the REST API request </exception>
+        /// <exception cref="AuthorizationException"> if there is any problem with  the REST API authorization (access token) </exception>
+        /// <exception cref="ResourceNotFoundException"> if the resource cannot be found </exception>
+        /// <exception cref="ServiceUnavailableException"> if the REST API service is not available (possibly due to rate limiting) </exception>
+        /// <exception cref="SmartsheetException"> if there is any other error during the operation </exception>
+        public virtual Attachment AttachFile(long sheetId, long commentId, Stream stream, string fileName, string? contentType)
+        {
+            return AttachFile("sheets/" + sheetId + "/comments/" + commentId + "/attachments", stream, fileName, contentType);
+        }
+
+        /// <summary>
         /// <para>Attaches a URL to the Comment.</para>
         /// <para>It mirrors to the following Smartsheet REST API method:
         /// POST /sheets/{sheetId}/comments/{commentId}/attachments</para>
@@ -142,6 +166,20 @@ namespace Smartsheet.Api.Internal
             this.Smartsheet.HttpClient.ReleaseConnection();
 
             return attachment;
+        }
+
+        /// <summary>
+        /// Attach file from stream.
+        /// </summary>
+        /// <param name="path"> the url path </param>
+        /// <param name="stream"> the file stream </param>
+        /// <param name="fileName"> the file name </param>
+        /// <param name="contentType"> the content Type </param>
+        /// <returns> the attachment </returns>
+        /// <exception cref="SmartsheetException"> the Smartsheet exception </exception>
+        private Attachment AttachFile(string path, Stream stream, string fileName, string? contentType)
+        {
+            return AttachFileFromStream(path, stream, fileName, contentType);
         }
     }
 }